+++ src/Control/Lens/Action/Internal.hs
@@ -28,7 +28,7 @@ module Control.Lens.Action.Internal
 import Control.Applicative
 import Control.Applicative.Backwards
 import Control.Monad
-import Data.Functor.Bind
+import Data.Functor.Semimonad
 import Data.Functor.Contravariant
 import Data.Functor.Identity
 import Data.Profunctor.Unsafe
@@ -93,17 +93,17 @@ instance Monad m => Effective m r (Effect m r) where
   ineffective = getEffect
   {-# INLINE ineffective #-}
 
-instance (Apply m, Semigroup r) => Semigroup (Effect m r a) where
+instance (Semiapplicative m, Semigroup r) => Semigroup (Effect m r a) where
   Effect ma <> Effect mb = Effect (liftF2 (<>) ma mb)
   {-# INLINE (<>) #-}
 
-instance (Apply m, Monad m, Monoid r) => Monoid (Effect m r a) where
+instance (Semiapplicative m, Monad m, Monoid r) => Monoid (Effect m r a) where
   mempty = Effect (return mempty)
   {-# INLINE mempty #-}
   Effect ma `mappend` Effect mb = Effect (liftM2 mappend ma mb)
   {-# INLINE mappend #-}
 
-instance (Apply m, Semigroup r) => Apply (Effect m r) where
+instance (Semiapplicative m, Semigroup r) => Semiapplicative (Effect m r) where
   Effect ma <.> Effect mb = Effect (liftF2 (<>) ma mb)
   {-# INLINE (<.>) #-}